# admin/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Admin API
    description: The admin service definition.
    version: 0.0.1
paths:
    /api/address/create:
        post:
            tags:
                - Admin
            operationId: Admin_CreateAddress
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAddressReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AddressInfo'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/address/default:
        put:
            tags:
                - Admin
            operationId: Admin_DefaultAddress
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AddressReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CheckResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/address/delete:
        delete:
            tags:
                - Admin
            operationId: Admin_DeleteAddress
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: uid
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CheckResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/address/list/uid:
        get:
            tags:
                - Admin
            operationId: Admin_AddressListByUid
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListAddressReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/address/update:
        put:
            tags:
                - Admin
            operationId: Admin_UpdateAddress
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateAddressReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CheckResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/users/captcha:
        get:
            tags:
                - Admin
            operationId: Admin_Captcha
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CaptchaReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/users/detail:
        get:
            tags:
                - Admin
            operationId: Admin_Detail
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDetailResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/users/login:
        post:
            tags:
                - Admin
            operationId: Admin_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RegisterReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/users/register:
        post:
            tags:
                - Admin
            operationId: Admin_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RegisterReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RegisterReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        AddressInfo:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                mobile:
                    type: string
                Province:
                    type: string
                City:
                    type: string
                Districts:
                    type: string
                address:
                    type: string
                postCode:
                    type: string
                isDefault:
                    type: integer
                    format: int32
        AddressReq:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                uid:
                    type: integer
                    format: int64
        CaptchaReply:
            type: object
            properties:
                captchaId:
                    type: string
                picPath:
                    type: string
        CheckResponse:
            type: object
            properties:
                success:
                    type: boolean
        CreateAddressReq:
            type: object
            properties:
                uid:
                    type: integer
                    format: int64
                name:
                    type: string
                mobile:
                    type: string
                Province:
                    type: string
                City:
                    type: string
                Districts:
                    type: string
                address:
                    type: string
                postCode:
                    type: string
                isDefault:
                    type: integer
                    format: int32
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ListAddressReply:
            type: object
            properties:
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/AddressInfo'
        LoginReq:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        RegisterReply:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                mobile:
                    type: string
                username:
                    type: string
                token:
                    type: string
                expiredAt:
                    type: integer
                    format: int64
            description: Data returned by registration and login
        RegisterReq:
            type: object
            properties:
                mobile:
                    type: string
                username:
                    type: string
                password:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateAddressReq:
            type: object
            properties:
                uid:
                    type: integer
                    format: int64
                name:
                    type: string
                mobile:
                    type: string
                Province:
                    type: string
                City:
                    type: string
                Districts:
                    type: string
                address:
                    type: string
                postCode:
                    type: string
                isDefault:
                    type: integer
                    format: int32
                id:
                    type: integer
                    format: int64
        UserDetailResponse:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                mobile:
                    type: string
                nickName:
                    type: string
                birthday:
                    type: integer
                    format: int64
                gender:
                    type: string
                role:
                    type: integer
                    format: int32
            description: user Detail returned
tags:
    - name: Admin